requirements.txt
fastapi
sqlmodel
uvicorn
httpx
langchain
sentence-transformers
chromadb-client
pydantic-settings
